LPD-78457 you should not de-invite a user you can not see and you do not invited

+++ modules/apps/headless/headless-object/headless-object-api/src/main/java/com/liferay/headless/object/util/v1_0/CollaboratorUtil.java
@@ -16,11 +16,14 @@ import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.model.Ticket;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.model.UserGroup;
+import com.liferay.portal.kernel.security.auth.GuestOrUserUtil;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.TicketLocalService;
 import com.liferay.portal.kernel.service.UserGroupLocalService;
 import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.service.permission.UserPermissionUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.OrderByComparatorFactoryUtil;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -275,7 +278,7 @@ public class CollaboratorUtil {
 			String className, long classNameId, long classPK, long companyId,
 			String emailAddress, JSONFactory jsonFactory,
 			SharingEntryService sharingEntryService,
-			TicketLocalService ticketLocalService,
+			TicketLocalService ticketLocalService, long userId,
 			UserLocalService userLocalService)
 		throws Exception {
 
@@ -289,6 +292,12 @@ public class CollaboratorUtil {
 				0, 0, user.getUserId(), classNameId, classPK);
 
 			if (sharingEntry != null) {
+				if ((sharingEntry.getUserId() != userId) &&
+					!_hasViewPermission(user)) {
+
+					throw new NoSuchModelException();
+				}
+
 				sharingEntryService.deleteSharingEntry(sharingEntry);
 			}
 		}
@@ -643,6 +652,12 @@ public class CollaboratorUtil {
 		}
 
 		return user.getUserId();
+	}
+
+	private static boolean _hasViewPermission(User user) throws Exception {
+		return UserPermissionUtil.contains(
+			GuestOrUserUtil.getPermissionChecker(), user.getUserId(),
+			ActionKeys.VIEW);
 	}
 
 	private static void _validateEmailActionIds(String[] actionIds) {
